requirements.txt
python-dotenv
py-cord>=2.0.0
aiosqlite
python-slugify
